# perf(gen): use `ht.ParseForm` instead of `ParseForm` method