… http/client.go
… http/multipart.go
… http/multipart_file.go