```
name                time/op
ParseForm/Custom-4     289µs ± 3%
ParseForm/Std-4        562µs ± 3%

name                speed
ParseForm/Custom-4  42.5MB/s ± 3%
ParseForm/Std-4     21.9MB/s ± 3%

name                alloc/op
ParseForm/Custom-4     231kB ± 0%
ParseForm/Std-4        496kB ± 0%

name                allocs/op
ParseForm/Custom-4     1.06k ± 0%
ParseForm/Std-4        2.12k ± 0%
```